v0.0.1
There are two sets of interest calculation tests:  Simple Daily
Interest and Compound Daily Interest.
The Compound Daily Interest set is turned off.  Please turn it on if
that is the calculation you are looking for.
Make sure interestFunc is set to self._cal_Compound_AccruedInt in
interestEarned method in Account class file.